pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fla>=0.6.3",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.32.0",
    "gunicorn>=23.0.0",
    "numpy>=1.26.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "sift-stack-py>=0.8.4",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "opencv-python>=4.11.0.86",
    "mediapipe>=0.10.21",
]
